Add ZeroMQ support, extend CAN suport, huge refactoring, bugfixes (#25)
- Brings ZeroMQ for communication between processes
- Add CAN bus server for centralized message sending
- Add CAN bus publisher for centralize incoming message notification for many clients
- Refactor the repository:
    - brain_service_backend -> brain_service
    - brain_service_frontend -> brain_service_display
    - brain_service_common is removed
    - Large code refactor inside the packages
- Add set of zakhar-specific functions to zakhar_profile.sh 
- Update Readme and docs to the common repository style
- brain_service_display: Remove error indicators with a status screen
- brain_service: Fix wifi id

Closes https://github.com/Zakhar-the-Robot/brain_pycore/issues/10, Closes #24, Closes #23, Closes #22, Closes #21, Closes #16, Closes #4, Closes #7

+++ scripts/services/brain_service_display_install.sh
@@ -10,7 +10,6 @@
 # e-mail:  mail@agramakov.me
 #
 # *************************************************************************
-
 if [[ $EUID -ne 0 ]]; then
    echo "This script must be run as root" 
    exit 1
@@ -20,14 +19,30 @@ SCRIPT_ROOT=$(dirname $(readlink -f "$0"))
 SCRIPT_NAME=$(basename "$0")
 function log { echo "- $1 [$(basename "$0")]" ;}
 # ----------------------------------------------------------------------------
+log "Install brain_service_display.service"
 
-SERVICE_FILE_NAME="brain.service"
+SERVICE_FILE_NAME="brain_service_display.service"
+SERVICE_SRC_FILE_PATH="$SCRIPT_ROOT/$SERVICE_FILE_NAME"
+SERVICE_DSC_FILE_PATH="/lib/systemd/system/$SERVICE_FILE_NAME"
 
-log "Stopping the service: $SERVICE_FILE_NAME"
-systemctl stop $SERVICE_FILE_NAME
 
-log "Reload: $SERVICE_FILE_NAME"
+if [ ! -f $SERVICE_DSC_FILE_PATH ]; then
+    log "Creating a $SERVICE_FILE_NAME file"
+else
+    log "$SERVICE_FILE_NAME exists! The service will be updated"
+    log "Stopping the service: $SERVICE_FILE_NAME"
+    systemctl stop $SERVICE_FILE_NAME
+    rm $SERVICE_DSC_FILE_PATH
+fi
+
+log "Writing the service"
+cp "$SERVICE_SRC_FILE_PATH" "$SERVICE_DSC_FILE_PATH"
+
+log "Reloading systemctl daemons"
 systemctl daemon-reload
+
+log "Enabling the service"
+systemctl enable $SERVICE_FILE_NAME
 
 log "Starting the service: $SERVICE_FILE_NAME"
 systemctl start $SERVICE_FILE_NAME
